# openapi/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ubu_orchestrator",
    "description": "Local async orchestration service for UbU Phase 1 dogfooding.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/task/{task_id}/decompose": {
      "post": {
        "tags": [
          "crate::api::user_action"
        ],
        "operationId": "decompose",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/task/{task_id}/done": {
      "post": {
        "tags": [
          "crate::api::user_action"
        ],
        "operationId": "done",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/task/{task_id}/reject": {
      "post": {
        "tags": [
          "crate::api::user_action"
        ],
        "operationId": "reject",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/task/{task_id}/snooze": {
      "post": {
        "tags": [
          "crate::api::user_action"
        ],
        "operationId": "snooze",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/task/{task_id}/start": {
      "post": {
        "tags": [
          "crate::api::user_action"
        ],
        "operationId": "start",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntryResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "LogEntryResponse": {
        "type": "object",
        "required": [
          "logId",
          "taskId",
          "action",
          "authoritySource"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/TaskActionKind"
          },
          "authoritySource": {
            "type": "string"
          },
          "logId": {
            "type": "string"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "taskId": {
            "type": "string"
          }
        }
      },
...
      "TaskActionKind": {
        "type": "string",
        "enum": [
          "start",
          "done",
          "snooze",
          "reject",
          "decompose"
        ]
      },
...
      "UserActionRequest": {
        "type": "object",
        "properties": {
          "note": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    }
  },
...
}