# src/api/user_action.rs
use axum::extract::{Path, State};
use axum::Json;
use serde::{Deserialize, Serialize};
use utoipa::ToSchema;

use crate::errors::Result;
use crate::services::log_service;
use crate::state::AppState;

#[derive(Debug, Clone, Copy, Serialize, Deserialize, ToSchema)]
#[serde(rename_all = "snake_case")]
pub enum TaskActionKind {
    Start,
    Done,
    Snooze,
    Reject,
    Decompose,
}

#[derive(Debug, Deserialize, ToSchema)]
#[serde(rename_all = "camelCase")]
pub struct UserActionRequest {
    #[serde(default)]
    pub note: Option<String>,
}

#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
#[serde(rename_all = "camelCase")]
pub struct LogEntryResponse {
    pub log_id: String,
    pub task_id: String,
    pub action: TaskActionKind,
    pub authority_source: String,
    pub note: Option<String>,
}

#[utoipa::path(
    post,
    path = "/task/{task_id}/start",
    params(("task_id" = String, Path)),
    request_body = UserActionRequest,
    responses((status = 200, body = LogEntryResponse))
)]
pub async fn start(
    State(state): State<AppState>,
    Path(task_id): Path<String>,
    Json(request): Json<UserActionRequest>,
) -> Result<Json<LogEntryResponse>> {
    Ok(Json(
        log_service::append_action(state, task_id, TaskActionKind::Start, request).await?,
    ))
}

#[utoipa::path(
    post,
    path = "/task/{task_id}/done",
    params(("task_id" = String, Path)),
    request_body = UserActionRequest,
    responses((status = 200, body = LogEntryResponse))
)]
pub async fn done(
    State(state): State<AppState>,
    Path(task_id): Path<String>,
    Json(request): Json<UserActionRequest>,
) -> Result<Json<LogEntryResponse>> {
    Ok(Json(
        log_service::append_action(state, task_id, TaskActionKind::Done, request).await?,
    ))
}

#[utoipa::path(
    post,
    path = "/task/{task_id}/snooze",
    params(("task_id" = String, Path)),
    request_body = UserActionRequest,
    responses((status = 200, body = LogEntryResponse))
)]
pub async fn snooze(
    State(state): State<AppState>,
    Path(task_id): Path<String>,
    Json(request): Json<UserActionRequest>,
) -> Result<Json<LogEntryResponse>> {
    Ok(Json(
        log_service::append_action(state, task_id, TaskActionKind::Snooze, request).await?,
    ))
}

#[utoipa::path(
    post,
    path = "/task/{task_id}/reject",
    params(("task_id" = String, Path)),
    request_body = UserActionRequest,
    responses((status = 200, body = LogEntryResponse))
)]
pub async fn reject(
    State(state): State<AppState>,
    Path(task_id): Path<String>,
    Json(request): Json<UserActionRequest>,
) -> Result<Json<LogEntryResponse>> {
    Ok(Json(
        log_service::append_action(state, task_id, TaskActionKind::Reject, request).await?,
    ))
}

#[utoipa::path(
    post,
    path = "/task/{task_id}/decompose",
    params(("task_id" = String, Path)),
    request_body = UserActionRequest,
    responses((status = 200, body = LogEntryResponse))
)]
pub async fn decompose(
    State(state): State<AppState>,
    Path(task_id): Path<String>,
    Json(request): Json<UserActionRequest>,
) -> Result<Json<LogEntryResponse>> {
    Ok(Json(
        log_service::append_action(state, task_id, TaskActionKind::Decompose, request).await?,
    ))
}